refactor: move request/response dto to `dtos` folder
