+++ src/sp-proxy/src/interface-adapters/delivery/CreateRemoteIdpController.ts
@@ -1,11 +1,11 @@
 import { IController } from '@sp-proxy/interface-adapters/protocols/IController'
-import { ICreateRemoteIdpRequest } from '@sp-proxy/interface-adapters/protocols/ICreateRemoteIdpRequest'
 import { IDeliveryMapper } from '@sp-proxy/interface-adapters/protocols/IDeliveryMapper'
 import { IRequest } from '@sp-proxy/interface-adapters/protocols/IRequest'
 import { IValidator } from '@sp-proxy/interface-adapters/protocols/IValidator'
 import { InputBoundary } from '@sp-proxy/use-cases/ports/InputBoundary'
 import { IRequestModel } from '@sp-proxy/use-cases/io-models/IRequestModel'
 import { AddRemoteIdpUseCaseParams } from '@sp-proxy/use-cases/io-models/RemoteIdpUseCaseParams'
+import { ICreateRemoteIdpRequest } from '@sp-proxy/interface-adapters/delivery/dtos/ICreateRemoteIdpRequest'
 
 export class CreateRemoteIdpController implements IController {
   constructor(
