+++ src/sp-proxy/src/interface-adapters/delivery/CreateRemoteIdpPresenter.spec.ts
@@ -1,9 +1,9 @@
 import { CreateRemoteIdpPresenter } from '@sp-proxy/interface-adapters/delivery/CreateRemoteIdpPresenter'
-import { ICreateRemoteIdpResponse } from '@sp-proxy/interface-adapters/protocols/ICreateRemoteIdpResponse'
 import { IResponse } from '@sp-proxy/interface-adapters/protocols/IResponse'
 import { SuccessResponseUseCaseParams } from '@sp-proxy/use-cases/io-models/SuccessResponseUseCaseParams'
 import { IResponseModel } from '@sp-proxy/use-cases/io-models/IResponseModel'
 import { EventEmitter } from 'stream'
+import { ICreateRemoteIdpResponse } from '@sp-proxy/interface-adapters/delivery/dtos/ICreateRemoteIdpResponse'
 
 interface SutTypes {
   sut: CreateRemoteIdpPresenter
