+++ src/sp-proxy/src/interface-adapters/delivery/GetRemoteIdpPresenter.ts
@@ -1,10 +1,10 @@
 import { IDeliveryMapper } from '@sp-proxy/interface-adapters/protocols/IDeliveryMapper'
 import { IResponse } from '@sp-proxy/interface-adapters/protocols/IResponse'
-import { RemoteIdpDeliveryProps } from '@sp-proxy/interface-adapters/protocols/RemoteIdpDeliveryProps'
 import { OutputBoundary } from '@sp-proxy/use-cases/ports/OutputBoundary'
 import { IResponseModel } from '@sp-proxy/use-cases/io-models/IResponseModel'
 import { RemoteIdpUseCaseParams } from '@sp-proxy/use-cases/io-models/RemoteIdpUseCaseParams'
 import { EventEmitter } from 'stream'
+import { RemoteIdpDeliveryProps } from '@sp-proxy/interface-adapters/delivery/dtos/RemoteIdpDeliveryProps'
 
 export class GetRemoteIdpPresenter
   implements OutputBoundary<IResponseModel<RemoteIdpUseCaseParams>>
