+++ src/sp-proxy/src/interface-adapters/delivery/dtos/GetByIdDTO.ts
