+++ src/sp-proxy/src/interface-adapters/delivery/dtos/IAddTrFromMetadataRequest.ts
