+++ src/sp-proxy/src/interface-adapters/delivery/dtos/IAddTrFromMetadataResponse.ts
