+++ src/sp-proxy/src/interface-adapters/delivery/mocks/fakeCreateRemoteIdpRequest.mock.ts
@@ -1,4 +1,4 @@
-import { ICreateRemoteIdpRequest } from '@sp-proxy/interface-adapters/protocols/ICreateRemoteIdpRequest'
+import { ICreateRemoteIdpRequest } from '@sp-proxy/interface-adapters/delivery/dtos/ICreateRemoteIdpRequest'
 import { IRequest } from '@sp-proxy/interface-adapters/protocols/IRequest'
 
 export const fakeCreateRemoteIdpRequest: IRequest<ICreateRemoteIdpRequest> = {
