+++ src/sp-proxy/src/interface-adapters/delivery/validators/AddTrFromMetadataValidator.ts
@@ -1,5 +1,5 @@
 import { isValidUrl } from '@sp-proxy/interface-adapters/delivery/validators/singles/isValidUrl'
-import { IAddTrFromMetadataRequest } from '@sp-proxy/interface-adapters/protocols/IAddTrFromMetadataRequest'
+import { IAddTrFromMetadataRequest } from '@sp-proxy/interface-adapters/delivery/dtos/IAddTrFromMetadataRequest'
 import { IRequest } from '@sp-proxy/interface-adapters/protocols/IRequest'
 import { IValidator } from '@sp-proxy/interface-adapters/protocols/IValidator'
