+++ src/sp-proxy/src/interface-adapters/protocols/IAddTrFromMetadataFacade.ts
@@ -1,4 +1,4 @@
-import { IAddTrFromMetadataRequest } from '@sp-proxy/interface-adapters/protocols/IAddTrFromMetadataRequest'
+import { IAddTrFromMetadataRequest } from '@sp-proxy/interface-adapters/delivery/dtos/IAddTrFromMetadataRequest'
 import { SuccessResponseUseCaseParams } from '@sp-proxy/use-cases/io-models/SuccessResponseUseCaseParams'
 
 export interface IAddTrFromMetadataFacade {
